# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Movies API",
    "description": "API to access data from the movies database",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "The local server"
    }
  ],
  "paths": {
...
    "/users/all": {
      "get": {
        "summary": "Returns all user data",
        "security": [
          {
            "BasicAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "required": false,
            "description": "the current page/set of data"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "required": false,
            "description": "the amount of data shown"
          }
        ],
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "User is unauthorized or not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "put": {
        "summary": "Updates user data",
        "security": [
          {
            "BasicAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "the id of the user"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "new@mail.com"
                  },
                  "gender": {
                    "type": "string",
                    "example": "Female"
                  },
                  "password": {
                    "type": "string",
                    "example": "new_password"
                  },
                  "role": {
                    "type": "string",
                    "example": "Project Manager"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "responses": {
          "201": {
            "description": "User data updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User Data Updated"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "User is unauthorized or not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Updates user data",
        "security": [
          {
            "BasicAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "the id of the user"
          }
        ],
        "tags": [
          "Users"
        ],
        "responses": {
          "201": {
            "description": "User data deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User Removed"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "User is unauthorized or not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "required": [
          "email",
          "gender",
          "password",
          "role"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "(supposed to be auto-generated) id of the user"
          },
          "email": {
            "type": "string",
            "description": "email of the user"
          },
          "gender": {
            "type": "string",
            "description": "the gender of the user"
          },
          "password": {
            "type": "string",
            "description": "the private password of the user"
          },
          "role": {
            "type": "string",
            "description": "the designated role of the user"
          }
        },
        "example": {
          "id": 101,
          "email": "example@email.com",
          "gender": "Male",
          "password": "secretpassword123",
          "role": "Construction Worker"
        }
      },
...
      "NotFound": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Error Not Found"
          }
        }
      },
      "ServerError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Internal Server Error"
          }
        }
      }
    }
  }
}